# src/main/java/com/vinodh/security/jwt/dto/LoginRequest.java
package com.vinodh.security.jwt.dto;



public class LoginRequest {
    private String email;
    private String password;

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }
}