# src/main/resources/static/openapi.yml
openapi: 3.0.3
info:
  title: Authentication API
  version: 1.0.0
  description: API endpoints for user registration and login

servers:
  - url: http://localhost:8080
    description: Local server

paths:
  /api/v1/auth/register:
    post:
      summary: Register a new user
      description: Register a new user in the system. The role defaults to "USER" if not provided.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              firstName: "Vinodh"
              lastName: "Sangavaram"
              email: "vinodh.sangavaram@vcti.com"
              password: "securePass123"
              role: "USER"
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  firstName:
                    type: string
                    example: "Vinodh"
                  lastName:
                    type: string
                    example: "Sangavaram"
                  email:
                    type: string
                    example: "vinodh.sangavaram@vcti.com"
                  role:
                    type: string
                    example: "USER"

  /api/v1/auth/login:
    post:
      summary: Login
      description: Authenticate a user and return a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "vinodh.sangavaram@vcti.com"
              password: "securePass123"
      responses:
        '200':
          description: JWT token returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

components:
  schemas:
    RegisterRequest:
      type: object
      properties:
        firstName:
          type: string
          example: "Vinodh"
        lastName:
          type: string
          example: "Sangavaram"
        email:
          type: string
          format: email
          example: "vinodh.sangavaram@vcti.com"
        password:
          type: string
          format: password
          example: "securePass123"
        role:
          type: string
          default: "USER"
          example: "USER"
      required:
        - firstName
        - lastName
        - email
        - password

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "vinodh.sangavaram@vcti.com"
        password:
          type: string
          format: password
          example: "securePass123"
      required:
        - email
        - password